+++ client/projects/cms/src/app/modules/dashboard/modules/module-instance/pages/instance-single/instance-single.component.ts
@@ -224,7 +224,12 @@ export class InstanceSingleComponent implements OnInit {
   back() {
     this.initialValue = '';
     this.currentValue = '';
-    this.router.navigate(['../..', 'overview'], {relativeTo: this.activatedRoute});
+
+    if (this.activatedRoute.snapshot.queryParams.back) {
+      this.router.navigate([this.activatedRoute.snapshot.queryParams.back]);
+    } else {
+      this.router.navigate(['../..', 'overview'], {relativeTo: this.activatedRoute});
+    }
   }
 
   valueChange(data: any, instance: Instance) {