+++ client/projects/cms/src/app/shared/modules/fb/form-builder-shared.module.ts
@@ -4,7 +4,7 @@ import {AngularFireStorage} from '@angular/fire/storage';
 import {MatButtonModule} from '@angular/material/button';
 import {MatIconModule} from '@angular/material/icon';
 import {
-  CUSTOM_COMPONENTS,
+  CUSTOM_COMPONENTS, CUSTOM_FIELDS,
   DbService as FDbService,
   FormBuilderModule,
   ROLE,
@@ -59,6 +59,10 @@ export function roleFactory(state: StateService) {
       useValue: {
         duplicate: DuplicateComponent
       }
+    },
+    {
+      provide: CUSTOM_FIELDS,
+      useValue: {}
     }
   ],
   declarations: [DuplicateComponent]